+++ packages/server/lib/controllers/v1/environment/postEnvironment.ts
@@ -45,7 +45,7 @@ export const postEnvironment = asyncWrapper<PostEnvironment>(async (req, res) =>
             res.status(400).send({
                 error: {
                     code: 'resource_capped',
-                    message: plan.name === 'free' ? 'Creating environment is only available for paying customer' : "Can't create more environments. "
+                    message: "Can't create more environments. "
                 }
             });
             return;
