+++ packages/types/lib/meta/api.ts
@@ -1,5 +1,6 @@
 import type { ApiError, Endpoint } from '../api';
 import type { DBEnvironment } from '../environment/db';
+import type { ApiPlan } from '../plans/http.api';
 
 export type GetMeta = Endpoint<{
     Method: 'GET';
@@ -8,6 +9,7 @@ export type GetMeta = Endpoint<{
     Error: ApiError<'user_not_found'>;
     Success: {
         data: {
+            plan: ApiPlan | null;
             environments: Pick<DBEnvironment, 'name'>[];
             version: string;
             baseUrl: string;
